Fixed issue #4 (circles' size)
Thanks to EriFa for pointing out this. Haplotype frequency (represented by the size of circles) was scaled by the radius of the circle instead of its area. Now, there is a standard area for a circle representing one individual (standardRadius^2 * PI) which is then multiplied by each haplotype absolute frequency to compute the scaled area. This is then divided by PI and square rooted to obtain the respective radius (as expected, d3.js uses radius to draw circles)